refactor: BangSearchScreen component 폴더 생성, 삽입
