+++ src/screens/bang/BangSearchScreen.styled.ts
