+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
-    "extends": "expo/tsconfig.base",
-    "compilerOptions": {
-        "jsx": "react",
-        "strict": true
-    }
+  "extends": "expo/tsconfig.base",
+  "compilerOptions": {
+    "jsx": "react",
+    "strict": true
+  }
 }
